Organise functionality out of Form1.cs
I've moved most of the functionality that doesn't directly interact with the UI out of Form1.cs. Updated the code to use FileStreams instead of loading the whole file into memory. Converted all the modifications so that they fit under an interface. Moved some misc functions into a Utilities class.